git-svn-id: svn://svn.code.sf.net/p/win32-framework/code@2758 40094c20-7226-0410-9e7a-ca24b3fb783e

+++ samples/ScintillaDemo/src/scintillaview.cpp
@@ -293,7 +293,7 @@ void ScintillaView::QuickPrint(LPCTSTR docName)
     di.lpszOutput = NULL;   // Do not print to file.
     printerDC.StartDoc(&di);
 
-    int maxPages = CollatePages();
+    int maxPages = static_cast<int>(CollatePages());
     for (int page = 0; page < maxPages; ++page)
     {
         // Start the page.
